requirements.txt
requests
Pillow
plexapi
python-dotenv